Added Calendar
Began the development of the calendar. First step is to make it look
nice. Get the scroll bar working well. Then integrate with json to make
the forwar and back buttons work.

Fixed the header so that it scrolls with you. Added buttons to go
forward and backward a week.