report.html update and internals.html
what I had written for the report was not nearly as much as we need, I
realize. Feel free to add your comments/things you want to talk about.
I might put them in my own words to keep it in one voice; it might not
matter though so add your own thoughts!